DRF/awesome27/requirements.txt
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
bcrypt==3.2.0
ipython==8.0.1
mycli==1.24.3
PyJWT==2.3.0
PyMySQL==1.0.2